Slowed down the emulation some. Will need to refactor later

+++ CHIP8-Emulator/constants.h
@@ -7,4 +7,4 @@ const int CH8_STACK_SIZE = 24;
 const int CH8_MESIZE = 0x1000;
 
 // Constants relating to clock rate and frames per second
-const float TIMER_RATE = 1000.0 / 60.0;
+const double TIMER_RATE = 1000.0 / 60.0;
